Fix Unicode support in hgettext utility
Previously, the hgettext utility used `show` function
to escape strings in the generated PO file. This lead to
Unicode characters being incorrectly escaped, for example,
… became \8230 instead of \x2026.

This tripped up xgettext command, especially when used with
--join-existing option.

According to gettext documentation [1], the strings use C syntax
so we are using functions for escaping C strings.

[1]: https://www.gnu.org/software/gettext/manual/html_node/PO-Files.html