Add bindTextDomainCodeset function
The bindTextDomainCodeset allows the user to set a specific codeset for
a given domain. This may be useful on systems, the where UTF-8 is not
set as the default codeset but the UTF-8 encoding of gettext messages
is needed.
One such example is having translated GUI labels in GTK4 on Windows,
where GTK4 requires UTF-8 strings, but Windows is set to use
region-specific codeset instead.